Put the commit ID in the version string
Finally, we are able to identify the builds more accurately. Now we check
if commit-id (cogito tool) is available and pass the output to vernum.c
compilation which then exports it as a build_id string.

The first line of --version is now a bit long, we should probably move the
build date to a second indented line or something.