Use $@ to simplify a rule and try to nullify metacharacters
I changed the %.gmo rule in po/Makefile to use $@ instead of a shell
variable, and made various tweaks in it to better support strange file
names, e.g. spaces in $(srcdir).  I suppose other makefiles in ELinks and
other rules in po/Makefile do not support those, but one must start
somewhere.  I don't think it will ever be feasible to support backslashes
and quotes without horrible contortions, but spaces and dashes should be
doable.

I also quoted the ~ in --check-accelerators=~ in case some shell might want
to expand it as the home directory. Bash 3.00.16(1) doesn't because the
word contains dashes.