Make elinks.conf(5) also be build from .txt files
This simplifies help2doc since it now just have to do one single thing with
output of either --config-help or --long-help. The new manpage is not as
fancy wrt. indentation and sub titles but it is more uniform and we get a
HTML version.

Generate options-{config,command}.txt with help2doc. Only conditionally
check for API header files, to speed up execusion.