CONFIG: Fix GC checking on Debian-based systems
This changes GC checking to use the EL_CHECK_OPTIONAL_LIBRARY macro to
reduce redundancy. This means that there is --with-gc instead of
--enable-gc so you can pass it a path, which is needed on Debian systems
where <gc.h> resides in /usr/include/gc/gc.h.

The macro is adapted accordingly to serve the above purpose. That is,
"$withval/include" and "$withval/lib" is only appended to CFLAGS and LIBS
if they exist (else "$withval" is appended to CFLAGS).