doc: Misc spelling and AsciiDoc compatibility fixes.
This part of dev-intro.txt doesn't work on AsciiDoc 7.1.2:

.Overview of the hierarchy of the various subsystems. At the bottom are \
 subsystems that provide functionality used by the upper layers.

AsciiDoc treats only the first line as the title and includes the
backslash in the XHTML output.  It looks like the only way to fix
dev-intro.txt is to merge the lines into one, but this would both make
the source ugly and somehow generate "Example:" at the beginning of
the title.  Because doc/Makefile does not currently run AsciiDoc on
dev-intro.txt, I'm leaving this part unchanged.