Define datarootdir in Makefile.config.in, for better Autoconf compatibility.
With Autoconf 2.60a, the default values of datadir, infodir, and
mandir refer to ${datarootdir}.  If Makefile.config.in does not define
datarootdir, Autoconf detects this and expands ${datarootdir} when it
substitutes expressions like @datadir@, but it also outputs the
following warning:

config.status: creating Makefile.config
config.status: WARNING:  /home/Kalle/src/elinks/Makefile.config.in seems to ignore the --datarootdir setting

According to a comment in config.status, "This hack should be removed
a few years after 2.60."  So it seems best to prepare for that now by
defining datarootdir = @datarootdir@ in Makefile.config.in.  Earlier
versions of Autoconf may leave that line unexpanded; but because the
makefiles do not directly refer to ${datarootdir}, there's no harm.