Bug 788: Define n_a variables if the POPpx macro of Perl needs them.
The configure script checks whether it is possible to compile a use of
POPpx without an n_a variable; if not, the source code then defines
those variables.  This is slower than including Perl's patchlevel.h
and comparing the version numbers to 5.8.8 but I expect this to be
more reliable as well.

[ Backported from commit 432969a3a96790f92ea0097795a36cfbf264999d in
  ELinks 0.12.GIT.  --KON ]