menu_handler: Don't call get_parent_ptr for redraw events
Only call get_parent_ptr for resize events so that the menu isn't moved
to the cursor position for normal redraw events.

Without this change, if ui.leds.enable and ui.clock.enable are 1 and
the user opens the tab menu, it first appears near the tab bar and
almost immediately jumps near the cursor.  With this change, the tab
menu stays near the tab bar.  However, if the user presses Ctrl+L,
then the menu still jumps; but that is less annoying than having it
jump on its own.

[ From commit 1a89589b13bf34e85a445ae7ba84602bc0522f81 in ELinks
  0.12.GIT.  --KON ]