Tell xgettext which functions expect format strings.
I constructed the list by grepping for "..." and looking for related
macros and va_list functions.  Also grepped for "*fmt", and "*format"
but the extra functions found that way (add_date_to_string,
format_command, subst_user_agent, etc.) handle format strings that
don't have the same syntax as in printf: in particular, type safety
does not depend on the order of format specifiers like it does in
printf.  Therefore, these format strings should not be subjected to
the "c-format" checks of msgfmt.