Tell xgettext which functions pass through format strings.
These are the same functions whose argument strings xgettext should
add to elinks.pot.  I also searched for uses of the functions that are
known to take format strings, in case the callers might take the
format string from the result of another function, but didn't find any
new ones.