Reset forms after reloading a document.
This is a fix for the bug from bugzilla, but I don't remember the bug number
now. Forms are reset when new documents are loaded or old ones are reloaded.
Because of gradual rerendering forms may be reset a few times
before a document is fully loaded, what may be annoying.