Revert 8bedf25b, "goto_url_hook takes 2 params not 1."
Revert 8bedf25bf84d60b143b16d0fe1b384dc17c93569.  Different versions
of ELinks call the Python goto_url_hook with different arguments:

- goto_url_hook(current) in 0.11.0, 0.11.1, and 0.11.2.

- goto_url_hook(new, current) in 0.11.2.GIT since 2007-11-27, commit
  8bedf25bf84d60b143b16d0fe1b384dc17c93569 which I'm now reverting.
  No released version of ELinks has used this variant.

- goto_url_hook(new) in 0.12.GIT since 2007-01-01, commit
  26473f72f59641aa60277f14f703f8a76dda5a82.  The hook can call
  elinks.current_url() if desired.

I think it will be less confusing to users if we drop the middle
variant and make the rest of the stable 0.11 branch call the hook in
the same way as 0.11.2 did.  The difference will then be only between
the 0.11 and 0.12 branches.