Check for gzclearerr in configure.  Document needing zlib >= 1.2.0.2.
It was reported at elinks-dev on 2007-06-03 that Solaris 10 comes with
zlib 1.1.4, which does not include gzclearerr(), which ELinks nowadays
requires.  It would be possible to rewrite the decompression support
to use deflate() directly and avoid stdio, in which case gzclearerr()
would not be needed.  That will take some time however, so I'm not
attempting it for ELinks 0.12.0.  Instead, I'm just disabling gzip
decompression entirely if zlib is too old.